backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart
python-dotenv
openai>=1.0.0
langchain>=0.1.0
requests
